add(blog): enhance blog update and delete functionality with author verification

+++ src/app/modules/blog/blog.service.ts
@@ -1,5 +1,6 @@
 import QueryBuilder from "../../builder/queryBuilder";
 import ApiError from "../../utils/ApiError";
+import { TUser } from "../user/user.interface";
 import { TBlog } from "./blog.interface";
 import { Blog } from "./blog.model";
 
@@ -19,14 +20,26 @@ const createBlog = async (blogContent: TBlog) => {
 	return result;
 };
 
-const updateBlog = async (blogId: string, updateContent: Partial<TBlog>) => {
+const updateBlog = async (
+	blogId: string,
+	updateContent: Partial<TBlog>,
+	userEmail: string
+) => {
 	// check blog already exist or not
 
-	const blog = await Blog.findById(blogId);
+	const blog = await Blog.findById(blogId).populate("author");
 
 	// if blog not exist throw error
 
 	if (!blog) throw new ApiError(404, "Blog not found");
+
+	// check if user is author of blog
+
+	const { email } = blog.author as Partial<TUser>;
+
+	if (userEmail !== email) {
+		throw new ApiError(401, "You are not author of this blog");
+	}
 
 	const result = await Blog.findByIdAndUpdate(blogId, updateContent, {
 		new: true,
@@ -39,14 +52,20 @@ const updateBlog = async (blogId: string, updateContent: Partial<TBlog>) => {
 	return result;
 };
 
-const deleteBlog = async (blogId: string) => {
+const deleteBlog = async (blogId: string, userEmail: string) => {
 	// check blog already exist or not
 
-	const blog = await Blog.findById(blogId);
+	const blog = await Blog.findById(blogId).populate("author");
 
 	// if blog not exist throw error
 
 	if (!blog) throw new ApiError(404, "Blog not found");
+
+	const { email } = blog.author as Partial<TUser>;
+
+	if (userEmail !== email) {
+		throw new ApiError(401, "You are not author of this blog");
+	}
 
 	const result = await Blog.findByIdAndDelete(blogId);
 
@@ -58,18 +77,18 @@ const deleteBlog = async (blogId: string) => {
 };
 
 const getAllBlogs = async (query: Record<string, unknown>) => {
-    const blogQuery = new QueryBuilder(Blog.find().populate("author"), query)
-        .search(['title','content'])
-        .filter()
-        .sort();
+	const blogQuery = new QueryBuilder(Blog.find().populate("author"), query)
+		.search(["title", "content"])
+		.filter()
+		.sort();
 
-    const response = await blogQuery.modelQuery;
+	const response = await blogQuery.modelQuery;
 
 	// check if response is empty
 
 	if (response.length <= 0) throw new ApiError(404, "No blogs found");
 
-    return response;
+	return response;
 };
 
 export const blogService = { createBlog, updateBlog, deleteBlog, getAllBlogs };
